# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/quests:
    get:
      operationId: QuestsController_list
      summary: List quests (filterable, paginated)
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            enum:
              - DRAFT
              - PUBLISHED
              - ARCHIVED
            type: string
        - name: difficulty
          required: false
          in: query
          schema:
            enum:
              - EASY
              - MEDIUM
              - HARD
              - EPIC
            type: string
        - name: categorySlug
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - quests
      security:
        - access-token: []
    post:
      operationId: QuestsController_create
      summary: Create a new quest (admin/moderator/business owner)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestDto"
      responses:
        "201":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/nearby:
    get:
      operationId: QuestsController_listNearby
      summary: List quests near coordinates (radius search)
      parameters:
        - name: lat
          required: false
          in: query
          schema:
            minimum: -90
            maximum: 90
            type: number
        - name: lng
          required: false
          in: query
          schema:
            minimum: -180
            maximum: 180
            type: number
        - name: radiusM
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100000
            default: 5000
            type: number
        - name: difficulty
          required: false
          in: query
          schema:
            enum:
              - EASY
              - MEDIUM
              - HARD
              - EPIC
            type: string
        - name: categorySlug
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/{id}:
    get:
      operationId: QuestsController_getById
      summary: Get quest by id
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
    patch:
      operationId: QuestsController_update
      summary: Update a quest (author or admin/moderator only)
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestDto"
      responses:
        "200":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
    delete:
      operationId: QuestsController_delete
      summary: Delete a quest (author or admin/moderator only)
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/{id}/start:
    post:
      operationId: QuestsController_start
      summary: Mark a quest as started for the current user
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/{id}/locations/{locationId}/check-in:
    post:
      operationId: QuestsController_checkInAtLocation
      summary: Check in at a quest waypoint (geolocation)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: locationId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestCheckInBodyDto"
      responses:
        "201":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/{id}/complete:
    post:
      operationId: QuestsController_complete
      summary: Mark a quest as completed for the current user
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
  /api/v1/quests/{id}/rate:
    post:
      operationId: QuestsController_rate
      summary: Rate a quest (upserts)
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateQuestDto"
      responses:
        "201":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
    delete:
      operationId: QuestsController_unrate
      summary: Remove your rating from a quest
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags:
        - quests
      security:
        - access-token: []
...
info:
  title: SideQuest API
  description: Gamified exploration platform - REST API
  version: "1"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the access token returned by /auth/login
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: number
        limit:
          type: number
        total:
          type: number
        totalPages:
          type: number
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    QuestLocationDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: object
        address:
          type: object
        latitude:
          type: number
        longitude:
          type: number
        radiusM:
          type: number
        orderIndex:
          type: number
        distanceM:
          type: number
          description: Distance in meters from the requested point
        isCompleted:
          type: boolean
          description: Set when the request carries a valid access token
      required:
        - id
        - latitude
        - longitude
        - radiusM
        - orderIndex
    ParticipantSummaryDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: object
        avatarUrl:
          type: object
        level:
          type: number
      required:
        - id
        - username
        - level
    QuestCategoryDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: object
        iconUrl:
          type: object
        coverImageUrl:
          type: object
        colorHex:
          type: object
        questCount:
          type: number
      required:
        - id
        - slug
        - name
        - questCount
    AchievementSummaryDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        iconUrl:
          type: object
        xpBonus:
          type: number
      required:
        - id
        - slug
        - name
        - xpBonus
    QuestRewardsDto:
      type: object
      properties:
        xp:
          type: number
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/AchievementSummaryDto"
      required:
        - xp
        - achievements
    QuestDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        summary:
          type: object
        description:
          type: string
        difficulty:
          type: string
          enum:
            - EASY
            - MEDIUM
            - HARD
            - EPIC
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
        xpReward:
          type: number
        estimatedDurationMin:
          type: object
        imageUrl:
          type: object
        coverImageUrl:
          type: object
        categoryId:
          type: object
        businessId:
          type: object
        authorId:
          type: object
        publishedAt:
          type: object
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        locations:
          type: array
          items:
            $ref: "#/components/schemas/QuestLocationDto"
        distanceM:
          type: number
          description: Distance in meters to the nearest location (nearby search)
        tags:
          type: array
          items:
            type: string
        rating:
          type: object
        ratingCount:
          type: number
        participantCount:
          type: number
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantSummaryDto"
        category:
          $ref: "#/components/schemas/QuestCategoryDto"
        rewards:
          $ref: "#/components/schemas/QuestRewardsDto"
      required:
        - id
        - slug
        - title
        - description
        - difficulty
        - status
        - xpReward
        - createdAt
        - updatedAt
        - tags
        - ratingCount
        - participantCount
        - rewards
    CreateQuestLocationDto:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radiusM:
          type: number
          default: 50
        orderIndex:
          type: number
          default: 0
      required:
        - latitude
        - longitude
    CreateQuestDto:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum:
            - EASY
            - MEDIUM
            - HARD
            - EPIC
          default: EASY
        xpReward:
          type: number
          default: 50
        estimatedDurationMin:
          type: number
        categoryId:
          type: string
        businessId:
          type: string
        imageUrl:
          type: string
          description: Thumbnail image URL
        coverImageUrl:
          type: string
          description: Hero / cover image URL
        locations:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestLocationDto"
      required:
        - slug
        - title
        - description
        - locations
    UpdateQuestDto:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum:
            - EASY
            - MEDIUM
            - HARD
            - EPIC
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
        xpReward:
          type: number
        estimatedDurationMin:
          type: number
        categoryId:
          type: string
        imageUrl:
          type: string
        coverImageUrl:
          type: string
        startsAt:
          type: string
        endsAt:
          type: string
    QuestCheckInBodyDto:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        accuracyM:
          type: number
          description: GPS accuracy in meters
      required:
        - latitude
        - longitude
    RateQuestDto:
      type: object
      properties:
        score:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - score
...